BDF2 works, now. Hopefully later too :)
We need to add higher order explicit methods because the beginning of highorder IMPLICIT needs to be done
with an explicit method with comparable order. This only needs to happen for implicit methods of order >=2
The current implicit classes availabe are:
backward euler, and bdf2

it is up to the user to decide which auxiliary stepper to pass to the implicit stepper.
But we will check at compile time that what is being passed is compatible stepper for beginning the simulation.

The process is a little convoluted on the user side because many things need to be created,
like polcies, steppers, etc. but this design should give us flexibility for the actual rom methods.