* adding multi_vector subpackage and specifically the epetra multivector. * Added test for it, minimal but it is there.
Note that a multivector can in general be seen as a multivector, or as a dense matrix.
However, the way the data is used in these two settings is different because they are different things.
We are going to support both, but the dense matrix class based on multivector still has to come (WIP).