Adding class for dense distributed matrix based on epetra. This is a follow up the previous commit, since also here we are using the epetra multivector BUT it is not wrapped into a core::Matrix object. This is to define a dense distributed matrix.
WARNING: what distinguishes the dense matrix wrapper from the actual multivector wrapper is
how we use the data for target operations.
For instance, when we need to do matrix-matrix multiplication for dense distributed, we need
to make sure that the type passed in is a dense matrix wrapper NOT a multivector wrapper.

**Multi-vector functions**:
These functions operate simultaneously but independently
on each vector in the multi-vector and produce individual results for each vector.

**Dense matrix functions**:
These functions operate on the multi-vector as a matrix.