adding mat-mat product for C = A * B where A = CRS epetra matrix B = densee matrix epetra
* The prod for C = A * B
when both A,B are dense is not working, so just put placeholder in there.
Trilinos does not havea native call for this. maybe I am wrong, but cannot find it...
Need to go back and fix it by adding a hardcoded one that I do myself.

* also added a core_container_distributed_trilinos_base.hpp
to be the base for some shared things among distributed containers in trilinos