Adding a draft function to do SVD on a trilinos CRS matrix. For now, this is just a function to do SVD, but we need to convert this into a class. So that we can choose methods, and tolerance, etc.
We are supporting doing svd on hermitian and nonhermitian matrices.
If the matrix is hermitian, i.e. symmetric for real values, then we simply do the call.
If matrix is not hermitian, then we call A*A^T and do svd on that.

I am adding two tests, comparing the results against precomputed results using python svd.