* Starting to remove all features from c++14 to make rompp only c++11 compliant. * things seem to be working ok.
We are NOT removing code that was there before, we are as much as possible try
to patch things so that when we can officially use c++14 whenever it is ok.

Currently things are working for gcc6.5.0 but we need to do more testing for other compilers.
It is really time to set up continous integration.

anyway, this is WIP. So stays in this branch until needed.