Changing order of trilinos libs to work for static/intel.
The original order did not have any issues for shared libs but was not working
on cee using intel compiler. So for the time being we are doing two thingsL
(a) changing the order to be same as tht used by sporc when linking static libs
(b) we add a few libs twice, i.e. we append at the bottom of the list.
Obviouisly, (b) is hack, but allows the build to work.