rom: adding new test to mimic sample mesh without mask (more below)
this effectively mimics what will be used when a real sample mesh is
used because we do not use the maskedLspg, but we use the
defaultlspg. the sample mesh is enabled within the computation of the
app rhs and jacobian, but rompp is totally obviously to it.
So the problme is run as if it was a regular problem, it just happens
that what the app provides is a ''reduced'' residual vector and jac matrix.