workaround to compile with cuda on cee machine
problem is that it seems current version of cuda on cee
does not compile a template class where class declares private friendship
with template parameter. Workaround is found at:

https://devtalk.nvidia.com/default/topic/1037721/nvcc-compilation-error-with-template-parameter-as-a-friend-within-a-namespace/

but we need to figure out if something is not right in the module, or if
the currrent cuda is really too old.