apps,tests: fix test pass regex for cmake detection
I made a mistake originally forgetting to add a pass regex for
cmake to detect that tests pass. So this came up when running tests with CUDA,
which flagged all apps tests to fail. But this was just a problem
with no having a proper regex. Now hopefully this will work as expected.