all: add version function to each package.
Tribits does not support header-only lib so adding a single
cc file with a version function in it allows us to leave all
functionalities in header files, but still allow
tribits to work properly.
In thoery, then, the libs are only needed if the
version function of one of the packages is called.
otherwise, a code using rompp as tpl should just use the headers.