add new op in core to do an update on a vector
Add new operation in core: core::ops::do_update
to use to update a vector. THis is useful because it
encapsulates/hides HOW we do it: if a vector type supports expression
templates, then we use that, but if it does not, like tpetra,
we use native functions, or we can do something else.
This way, we can write the operations inside ode steppers
in a single way and deletage implementations.